feat(kernels): move sink
