+++ crates/kernels/jstz_lightweight_kernel/Cargo.toml
@@ -30,6 +30,7 @@ tezos_crypto_rs.workspace = true
 jstz_mock = { path = "../../jstz_mock" }
 tezos-smart-rollup-mock.workspace = true
 serde_json.workspace = true
+jstz_utils = { path = "../../jstz_utils", features = ["test_utils"] }
 
 [features]
 lightweight-kernel = []
